requirements.txt
# DeepSeek-OCR 通用依赖
# 适用于 DGX Spark (ASUS GX10) + CUDA 13.0 + ARM64
#
# 注意：
# 1. PyTorch 需要单独安装，见 README.md
# 2. Transformers 和 Tokenizers 需要单独安装，不同环境版本不同，见 README.md

# 图像处理
Pillow>=10.0.0
einops>=0.8.0
numpy>=2.0.0

# 工具库
easydict
addict
tqdm

# PDF 处理
PyMuPDF
img2pdf
pdf2image
